Free version update:
@Sdwz	Sdwz	Memory of Disaster, part 1 done			d2e422f
@Sdwz	Sdwz	Fixed level 101 aura AVE and Talisman Insanity
skills			6cf40ed
@Sdwz	Sdwz	Fix item consume			313ba02
@Sdwz	Sdwz	Updating ensoul options with proper skill id, fixed a typo in
ensoul …  …			c40d713
@Sdwz	Sdwz	Changed nextAction support with now ATTACK and CAST			0be3031
@Sdwz	Sdwz	Fixing next action that should use current intention, and not
next			1e686e4
@Sdwz	Sdwz	Fixed typo in Cloud Breaker skill tree			5487e04
@Sdwz	Sdwz	skill typo			ba92361
@Sdwz	Sdwz	Some Ertheia skills change			8704839
@Sdwz	Sdwz	Lots of fixes for Ertheia skills			b73295c
@Sdwz	Sdwz	Adding ReduceDamage effect			d805d1c
@Sdwz	Sdwz	Removing item restriction			a36488e
@Sdwz	Sdwz	Adding AttackAttributeAdd effect			55c4e73
@Gladicek	Gladicek	Retail html typo in Q10358			934a236
@Sdwz	Sdwz	Fixing issue with alive npc in instance  …			aa75437
@Sdwz	Sdwz	Some Ertheia skills improvements + npc alive shit this
time			47c4b68
@Sdwz	Sdwz	Fixed TriggerSkillByDeathBlow with previous commit			8d61ea1
@Gladicek	Gladicek	Fixed QI ID for 3rd class transfer quest			1304d04
@Gladicek	Gladicek	GM Shop - Part 3			1255e4a
@Sdwz	Sdwz	Fixing ReduceHp NPE when in water or falling from high
height			697179a
@Sdwz	Sdwz	Fix SHOTS_BONUS usage, patch by @Nik--			0fef2da
@Sdwz	Sdwz	UseItem cleanup by @malyelfik			748dc71
@Sdwz	Sdwz	Memory of Disaster done			a9fb16f
@Sdwz	Sdwz	Fixing typo in ensoulFees			b53a51b
@Sdwz	Sdwz	Missing configuration key			281daed
@Sdwz	Sdwz	Eviscerator skills cleanup and correction			97eb7f3
@Gladicek	Gladicek	Updated htmls for FortressSiegeManager AI
…			b23be9f
@Sdwz	Sdwz	Massive cubic anti shit done :  …			f973265
@Sdwz	Sdwz	Retail skills for Buff cubic			34b0f33
@UnAfraid	UnAfraid	Addressing the NPE in PrivateStoreManageListSell:
…			39c1cc4
@ChaosPaladin	ChaosPaladin	Fixing trigger for skill 30509 Back
Step			a8bd5ac
@Sdwz	Sdwz	Changing DAX skills to handle FlyMove. Removed EnemyCharge,
FlyMove e…  …			8ec0ffe
@Sdwz	Sdwz	wtf you do here			4d60971
@ChaosPaladin	ChaosPaladin	Updating multisells Moe/Eeny. Implement new
chests			9099a51
@Sdwz	Sdwz	Fixing skill learn tooltip			004c1da
@Sdwz	Sdwz	Added AVE for Shillien Protection / Divinity of
Einhasad			fea0131
@Sdwz	Sdwz	Fix for race skills, which are learnt upon awakening and not
char create			79504b4
@Sdwz	Sdwz	Buff cubic template			2b0b760
@Sdwz	Sdwz	Fixing one shot hp bar update			cfe04c4
@Gladicek	Gladicek	Fix for Q10322			e78676d
@ChaosPaladin	ChaosPaladin	Fix bit in NpcBody			b5b059e
@UnAfraid	UnAfraid	Fixing up old weapons turned into EtcItem			3c99f23
@UnAfraid	UnAfraid	Restoring defaultAction of EtcItems			84f37bc
@UnAfraid	UnAfraid	Fixed applying SA options of non equipped items
…			4514f47
@UnAfraid	UnAfraid	Fixed Mark of Void to not cast the debuff on you, by
@Nik-			034e3e4
@ChaosPaladin	ChaosPaladin	Update Item Auction to Underground			91cd64a
@UnAfraid	UnAfraid	Spawns spawn methods			e7267ae
@UnAfraid	UnAfraid	More spawn related tweaks			e3b1ca2
@UnAfraid	UnAfraid	Undying Zone support 			d8ef8f4
@Sdwz	Sdwz	NPE fix in spawndata			655410a
@Sdwz	Sdwz	Fix for festival sweeper			f508ab9
@UnAfraid	UnAfraid	Some small optimizations requested by @Sdwz			779f7fa
@Sdwz	Sdwz	Getting ride of flyRadius custom parameter			1a38112
@ChaosPaladin	ChaosPaladin	Fix item count for learn skill
10956			e027dd0
@Gladicek	Gladicek	initialShortcuts:  …  			8151abd
@Sdwz	Sdwz	Skill enchanting support :  …			8e41465
@Gladicek	Gladicek	Fixed missing break in MonumentOfHeroes AI.
…			2a8ad31
@Sdwz	Sdwz	Fixed Beauty Shop			4cdf624
@Sdwz	Sdwz	Making sublevel to inherit from level value  …			a49a770
@Sdwz	Sdwz	Fixed the remaining brooch			f41a10f
@UnAfraid	UnAfraid	Updating Doors Emitter			09e99b6
@UnAfraid	UnAfraid	Fixing some typos in enchant system			783c474
@Sdwz	Sdwz	Fixing enchant fail value			b9e540a
@UnAfraid	UnAfraid	Adding some util methods to Instance class
…			b468c9f
@Sdwz	Sdwz	Dwarven Fighter skill tree correction			66241f0
@UnAfraid	UnAfraid	Fixed enchanting system route change 			9c03c85
@Gladicek	Gladicek	Reworked MusemDungeon instance 			9764dcb
@Gladicek	Gladicek	Some improvements by @Sdw for MuseumDungeon			302d91c
@Sdwz	Sdwz	Reworking a bit skillhashcode			ea15c1d
@Sdwz	Sdwz	Adding pDefMod to EnergyAttack and
PhysicalAttackWeaponBonus			e1e7305
@Sdwz	Sdwz	Adding sublevel in ExAbnormalStatusUpdateFromTarget			6d07f60
@Sdwz	Sdwz	Adding sub level in RequestDispel			5efdc21
@Sdwz	Sdwz	Fixing previous commit (I was drugged or what)			63f9166
@UnAfraid	UnAfraid	Adding general method for checking for the current
instance  …			48cbdf1
@UnAfraid	UnAfraid	AbstractInstance: Enforce template id/s to be
defined			968ddd0
@UnAfraid	UnAfraid	Reverting some Arrow's bodypart since they're also
considered EtcItem			b9c173e
@UnAfraid	UnAfraid	Fixing re-logging causing equipped items that doesn't
matches conditions  …			07ca41f
@UnAfraid	UnAfraid	Fixing up some FAN/SQUARE skills effecting targets
two times			4da6379
@Gladicek	Gladicek	Mini update for
Q10301_ShadowOfTerrorBlackishRedFog			65280ba
@Sdwz	Sdwz	Fixing fixing npc usage of a non existing skill			9c33cd5
@Sdwz	Sdwz	Making effect for sublevel fall back to level, thank Nos
!			d602c79
@Sdwz	Sdwz	Fixing enchantable status of skills. We now properly used
skill id an…  …			b4e454d
@Sdwz	Sdwz	Properly fixing brooch, issue reported by Mobius			abf232c
@Sdwz	Sdwz	Fixing issue with shield stats :  …			52aa623
@Sdwz	Sdwz	subLevel support for CallSkill			9b9c778
@Sdwz	Sdwz	Fixing improper break route enchant type maths			3310f36
@Sdwz	Sdwz	Target conditions support (as a working temporary way till we
find out			023051d
@Gladicek	Gladicek	Some changes O.o  …			4a46a8c
@Sdwz	Sdwz	Removing dual class revelation skills if subclass is
changed			2f66149
@Gladicek	Gladicek	Jeremy doesn't like event changes =(			77046fd
@ChaosPaladin	ChaosPaladin	Added enchant for Tyrrs/Sigels
skills			3e01abc
@Sdwz	Sdwz	Fixed auto soulshot			03557af
@Gladicek	Gladicek	Reverted commit for DoorInstance			8e463cb
@ChaosPaladin	ChaosPaladin	Added enchant for
Othells/Wynns/Iss/Feohs			ff40808
@Gladicek	Gladicek	Added AI for Ruins of Ye Sagira  …			45ebe21
@Gladicek	Gladicek	Eclipse... STAPH!			7d2c886
@ChaosPaladin	ChaosPaladin	Added enchant for Aero Healers
skill			76b07d4
@Sdwz	Sdwz	Fixing vampiric attack			f009e56
@ChaosPaladin	ChaosPaladin	Skills for PvP/PvE R99 items			ebf367b
@ChaosPaladin	ChaosPaladin	Fix bit in previous commit			12b5da6
@Sdwz	Sdwz	Helios initial support			0b093e1

+++ l2junity-gameserver/src/main/java/org/l2junity/gameserver/data/sql/impl/SummonSkillsTable.java
@@ -36,7 +36,7 @@ import org.slf4j.LoggerFactory;
 public class SummonSkillsTable
 {
 	private static Logger LOGGER = LoggerFactory.getLogger(SummonSkillsTable.class);
-	private final Map<Integer, Map<Integer, L2PetSkillLearn>> _skillTrees = new HashMap<>();
+	private final Map<Integer, Map<Long, L2PetSkillLearn>> _skillTrees = new HashMap<>();
 	
 	protected SummonSkillsTable()
 	{
@@ -54,7 +54,7 @@ public class SummonSkillsTable
 			while (rs.next())
 			{
 				final int npcId = rs.getInt("templateId");
-				Map<Integer, L2PetSkillLearn> skillTree = _skillTrees.get(npcId);
+				Map<Long, L2PetSkillLearn> skillTree = _skillTrees.get(npcId);
 				if (skillTree == null)
 				{
 					skillTree = new HashMap<>();
