+++ l2junity-gameserver/src/main/java/org/l2junity/gameserver/model/actor/instance/L2NpcInstance.java
@@ -140,7 +140,7 @@ public class L2NpcInstance extends Npc
 		
 		if (skills.isEmpty())
 		{
-			final Map<Integer, SkillLearn> skillTree = SkillTreesData.getInstance().getCompleteClassSkillTree(classId);
+			final Map<Long, SkillLearn> skillTree = SkillTreesData.getInstance().getCompleteClassSkillTree(classId);
 			final int minLevel = SkillTreesData.getInstance().getMinLevelForNewSkill(player, skillTree);
 			if (minLevel > 0)
 			{
